Fake "API" data source
There was no data source for our mock data.

Set up a fake API using the InMemoryWebApiModule, to mimic an actual
server that provided the data. Added the URL for that API to the shared
folder, so that client-side code can also use the same variable and
ensure our client is pointing at wherever our API actually is.