Partial wiring on submitting from Make a Transfer
The Submit button of the Make a Transfer component did nothing.

Now it correctly updates our data, deducting funds from the user and
adding a transaction to our log. Implementing this required caching the
data and manipulating it in our services, rather than using http. It
seems the in-memory-web-api module's handling of PUT and POST requests
is broken.